arm64: dts: rockchip: fix ethernet phy probe on rock-3a
The following was observed on my Radxa ROCK 3 Model A board:

  rk_gmac-dwmac fe010000.ethernet eth0: no phy at addr -1
  rk_gmac-dwmac fe010000.ethernet eth0: __stmmac_open: Cannot attach to PHY (error: -19)

This board has a RTL8211F PHY and the datasheet detail the following
regarding a PHY reset:

  For a complete PHY reset, this pin must be asserted low for at least
  10ms. Wait for a further 30ms before accessing the PHY register.

However, the Generic PHY driver does not handle PHY reset in this way.
It will assert the pin, probe and then deassert the pin.

The mdio bus is reset using an assert and deassert cycle and happens
just before the PHY probe. Take advantage of this and move the
reset-gpios from the phy node to the mdio bus node so that the PHY is
reset and ready in time for probe.

This also reduces the delay after deassert to closer match the timing
described in the datasheet.

After these changes the PHY is successfully discovered:

  rk_gmac-dwmac fe010000.ethernet eth0: PHY [stmmac-0:00] driver [RTL8211F Gigabit Ethernet] (irq=POLL)

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>